Add InitActivity that downloads DB before MainActivity
