+++ app/src/main/AndroidManifest.xml
@@ -15,18 +15,19 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity
-            android:name=".herd_member_details.HerdMemberDetails"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
-        <activity android:name=".DownloadDatabase" />
-        <activity
-            android:name=".main_activity.MainActivity"
-            android:windowSoftInputMode="adjustNothing"
-            android:theme="@style/AppTheme.NoActionBar">
+        <activity android:name=".InitActivity" android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".herd_member_details.HerdMemberDetails"
+            android:theme="@style/AppTheme.NoActionBar" />
+        <activity
+            android:name=".main_activity.MainActivity"
+            android:theme="@style/AppTheme.NoActionBar"
+            android:windowSoftInputMode="adjustNothing">
         </activity>
     </application>
