+++ app/src/main/res/layout/activity_init.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".InitActivity">
+
+    <TextView
+        android:id="@+id/initText"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:text="@string/initializing"
+        android:textAlignment="center"
+        android:gravity="center_vertical"
+        android:textSize="@dimen/font_h1"/>
+
+    <ProgressBar
+        android:id="@+id/progress_loader"
+        style="?android:attr/progressBarStyle"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_centerInParent="true"
+        android:visibility="visible" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
